Created env variable to change strapi url and made actions flexible
I had some actions hard coded initially to get the data so I could work on the layout.  Brian's headshot, which might be changed once or twice a year, should not be hard coded so I fixed that to fetch his headshot.url - also elminated corresponding code in the realtorReducer since it was overkill. Swapped out most of the hard coded localhost api calls with my environment variable.