Separated development database from production database
I should have heeded the warnings that you should not use the same database for development and production.  The relation _id changed from dev to prod and images would not work.  If I uploaded the image in dev, it would only appear in development, if I deleted the uploaded file and uploaded it again in prod, it would only appear in prod.