File cleanup
A bunch of file clean up.  Got rid of the great majority of commented out code (left a few as those are on my TODO list)  and got rid of console logs. Fixed the padding for the drop down menu - after a user logs in, the link container had a small gap on the right edge of the screen.