Updated listing model & added conditional to open house section
I changed garage field type from int to string since Brian needs to be able to add descriptions (e.g. "detached").  Added a check to see if the retrieved open_house date is > than current date - if it's not, then it is not displayed.  Unfortunately, Strapi does not allow you to clear the date once it has been set, you can change it, but not clear it, so this was the best solution at the moment to get rid of an open house that was set for example purposes.